+++ packages/components/src/pages/VsCodeExtension.vue
@@ -220,6 +220,22 @@
               <ReloadIcon class="control-button__icon" />
             </button>
           </v-popper>
+          <v-popper
+            class="hover-text-popper"
+            text="Toggle fullscreen"
+            placement="left"
+            text-align="left"
+            v-if="allowFullscreen"
+          >
+            <button
+              data-cy="fullscreen-button"
+              class="control-button"
+              :data-enabled="isFullscreen"
+              @click="toggleFullscreen"
+            >
+              <component :is="fullscreenIcon" class="control-button__stroke" />
+            </button>
+          </v-popper>
         </template>
       </v-tabs>
 
@@ -349,6 +365,8 @@ import DiagramGray from '@/assets/diagram-empty.svg';
 import StatsIcon from '@/assets/stats-icon.svg';
 import ExclamationIcon from '@/assets/exclamation-circle.svg';
 import ScissorsIcon from '@/assets/scissors-icon.svg';
+import FullscreenEnterIcon from '@/assets/fullscreen.svg';
+import FullscreenExitIcon from '@/assets/fullscreen-exit.svg';
 import VDetailsPanel from '../components/DetailsPanel.vue';
 import VDiagramComponent from '../components/DiagramComponent.vue';
 import VDiagramSequence from '../components/DiagramSequence.vue';
@@ -387,6 +405,7 @@ import {
   SET_SELECTED_SAVED_FILTER,
   SET_HIGHLIGHTED_EVENTS,
   SET_FOCUSED_EVENT,
+  SET_COLLAPSED_ACTION_STATE,
 } from '../store/vsCode';
 import isPrecomputedSequenceDiagram from '@/lib/isPrecomputedSequenceDiagram';
 import { SAVED_FILTERS_STORAGE_ID } from '../components/FilterMenu.vue';
@@ -421,6 +440,8 @@ export default {
     StatsIcon,
     ExclamationIcon,
     ScissorsIcon,
+    FullscreenEnterIcon,
+    FullscreenExitIcon,
   },
 
   store,
@@ -448,6 +469,7 @@ export default {
       maxSeqDiagramCollapseDepth: 9,
       sequenceDiagramDiffMode: false,
       isActive: true,
+      isFullscreen: false,
     };
   },
 
@@ -465,6 +487,10 @@ export default {
       default: () => [],
     },
     hideDetailsPanel: {
+      type: Boolean,
+      default: false,
+    },
+    allowFullscreen: {
       type: Boolean,
       default: false,
     },
@@ -821,6 +847,10 @@ export default {
     hasStats() {
       return this.stats && this.stats.functions && this.stats.functions.length > 0;
     },
+
+    fullscreenIcon() {
+      return this.isFullscreen ? FullscreenExitIcon : FullscreenEnterIcon;
+    },
   },
 
   methods: {
@@ -921,7 +951,7 @@ export default {
         state.expandedPackages = this.expandedPackages.map((expandedPackage) => expandedPackage.id);
 
       if (this.selectionStack.length > 1) {
-        state.selectedObjects = this.selectionStack.map(this.codeObjectToIdentifier);
+        state.selectedObjects = this.selectionStack.map(this.codeObjectToIdentifier.bind(this));
       } else if (this.selectionStack.length === 1) {
         state.selectedObject = this.codeObjectToIdentifier(this.selectedObject);
       } else if (this.selectedLabel) {
@@ -1081,6 +1111,7 @@ export default {
     },
 
     resetDiagram() {
+      this.$store.commit(SET_COLLAPSED_ACTION_STATE, []);
       this.seqDiagramCollapseDepth =
         DEFAULT_SEQ_DIAGRAM_COLLAPSE_DEPTH > this.maxSeqDiagramCollapseDepth
           ? this.maxSeqDiagramCollapseDepth
@@ -1311,6 +1342,43 @@ export default {
     },
     handleNewCollapseDepth(newDepth) {
       this.seqDiagramCollapseDepth = newDepth;
+    },
+    async enterFullscreen() {
+      const body = document.body;
+      const requestMethod =
+        body.requestFullScreen ||
+        body.webkitRequestFullScreen ||
+        body.mozRequestFullScreen ||
+        body.msRequestFullScreen;
+      if (requestMethod) {
+        try {
+          await requestMethod.call(body);
+        } catch (e) {
+          console.warn(e);
+        }
+      }
+    },
+    async exitFullscreen() {
+      const requestMethod =
+        document.exitFullscreen ||
+        document.webkitExitFullscreen ||
+        document.mozCancelFullScreen ||
+        document.msExitFullscreen;
+      if (requestMethod) {
+        try {
+          await requestMethod.call(document);
+        } catch (e) {
+          console.warn(e);
+        }
+      }
+    },
+    toggleFullscreen() {
+      if (this.isFullscreen) {
+        this.exitFullscreen();
+      } else {
+        this.enterFullscreen();
+      }
+      this.isFullscreen = !this.isFullscreen;
     },
   },
 
@@ -1590,6 +1658,17 @@ code {
         &:active {
           color: $gray5;
           transition-timing-function: ease-out;
+        }
+
+        svg path {
+          stroke: $lightgray2;
+        }
+
+        &__stroke {
+          width: 16px;
+          height: 14px;
+          stroke-width: 4;
+          fill: none;
         }
 
         &__icon {